added migrations for notices and balances tables, and changed migrations for operations, credits, checks tables, corrected related columns types according to column types of parent tables ids columns

+++ database/migrations/2021_05_31_190746_create_checks_table.php
@@ -19,7 +19,7 @@ class CreateChecksTable extends Migration
             $table->string('title_en')->nullable();
             $table->string('title_ru')->nullable();
             $table->double('sum');
-            $table->unsignedInteger('operation_id');
+            $table->unsignedBigInteger('operation_id');
             $table->timestamps();
         });
     }
